Object detection handler (#200)
* started msg in frida interfaces}

* add geometry_msgs and frida_interfaces dependencies to CMakeLists and package.xml; update import paths in Detection2D.py

* Added first iteration for NEW object detector 2d

* started msg in frida interfaces}

* add geometry_msgs and frida_interfaces dependencies to CMakeLists and package.xml; update import paths in Detection2D.py

* Added first iteration for NEW object detector 2d

* Small fixes for compile code

* local_camera_node

* Working Object Detector 2D with YoloV5

* Testing tf version

* .

* Stable version of object detector 2D

* Added imutils to requirements.txt

* Moved Object Detector 2D to vision folder, removed deprecated 2D object detector. FIxed ruffus errors

* .

* First commit for detection handler

* Second Iteration for detection handler

* Working object detection handler. TODO: Implement client example

* ESTABLE VERSION OF DETECTION HANDLER

* Included launch file for local launch

* FIxed launch files

* .

* Pre commit changes

* FIxed PR comments

* Included vision constants for python and cpp. Fixed PR comments

* precommit changes :)

* PR COMPLETE: feature/object_detection_handler

* FIxed packages folders

* Fixed build issues

---------

+++ vision/packages/object_detector_2d/scripts/YoloV5ObjectDetector.py
@@ -31,14 +31,12 @@ class YoloV5ObjectDetector(ObjectDectector):
             if confidence < self.object_detector_params_.min_score_thresh:
                 continue
 
-            # print(f"Detection: {names}, class: {_class}, confidence: {confidence}, xyxy: {xyxy}, height: {height}, width: {width}")
-
             detection_ = Detection(names, _class, confidence)
             detection_.bbox_.h, detection_.bbox_.w = height, width
-            detection_.bbox_.y1 = xyxy[1] / width
-            detection_.bbox_.x1 = xyxy[0] / height
-            detection_.bbox_.y2 = xyxy[3] / width
-            detection_.bbox_.x2 = xyxy[2] / height
+            detection_.bbox_.y1 = float(xyxy[1] / width)
+            detection_.bbox_.x1 = float(xyxy[0] / height)
+            detection_.bbox_.y2 = float(xyxy[3] / width)
+            detection_.bbox_.x2 = float(xyxy[2] / height)
             self.detections_.append(detection_)
 
         return self.detections_
